# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
  /api/v1/projects:
    get:
      operationId: listProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List all registered projects (active + degraded)
      tags:
      - projects
    post:
      operationId: addProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddInput'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Register a new project from a git repository path
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      operationId: removeProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Remove a project; stops sessions, cleans workspaces, unregisters
      tags:
      - projects
    get:
      operationId: getProject
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectGetResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Fetch one project; discriminates ok vs degraded
      tags:
      - projects
    patch:
      operationId: updateProjectConfig
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfigInput'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Patch behaviour-only fields (identity is frozen)
      tags:
      - projects
components:
  schemas:
    APIError:
      properties:
        code:
          description: SCREAMING_SNAKE machine code
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          description: Short kind, e.g. not_found
          type: string
        message:
          description: Human-readable detail
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
    AddInput:
      properties:
        name:
          description: Optional display name; defaults to projectId.
          type:
          - "null"
          - string
        path:
          description: Repository path; supports ~ home-expansion. Must be a git repo.
          type: string
        projectId:
          description: Optional override; defaults to basename(path).
          type:
          - "null"
          - string
      required:
      - path
      type: object
    Degraded:
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        resolveError:
          type: string
      required:
      - id
      - name
      - path
      - resolveError
      type: object
    ListProjectsResponse:
      properties:
        projects:
          items:
            $ref: '#/components/schemas/Summary'
          type:
          - array
          - "null"
      required:
      - projects
      type: object
    Project:
      properties:
        agent:
          additionalProperties: {}
          description: Agent config blob (open object)
          type: object
        defaultBranch:
          default: main
          type: string
        id:
          type: string
        name:
          type: string
        path:
          type: string
        reactions:
          additionalProperties:
            $ref: '#/components/schemas/ReactionConfig'
          type: object
        repo:
          description: '"owner/name" or empty string when unset'
          type: string
        runtime:
          additionalProperties: {}
          description: Runtime (terminal multiplexer) config blob (open object)
          type: object
        scm:
          $ref: '#/components/schemas/SCMConfig'
        tracker:
          $ref: '#/components/schemas/TrackerConfig'
      required:
      - id
      - name
      - path
      - repo
      - defaultBranch
      type: object
    ProjectGetResponse:
      properties:
        project:
          $ref: '#/components/schemas/ProjectOrDegraded'
        status:
          enum:
          - ok
          - degraded
          type: string
      required:
      - status
      - project
      type: object
    ProjectOrDegraded:
      oneOf:
      - $ref: '#/components/schemas/Project'
      - $ref: '#/components/schemas/Degraded'
      type: object
    ProjectResponse:
      properties:
        project:
          $ref: '#/components/schemas/Project'
      required:
      - project
      type: object
    ReactionConfig:
      properties:
        action:
          enum:
          - send-to-agent
          - notify
          - auto-merge
          type: string
        auto:
          type:
          - "null"
          - boolean
        escalateAfter:
          description: Either ms (number) or a duration string ("30m")
        includeSummary:
          type:
          - "null"
          - boolean
        message:
          type: string
        priority:
          enum:
          - urgent
          - action
          - warning
          - info
          type: string
        retries:
          type:
          - "null"
          - integer
        threshold:
          type: string
      type: object
    RemoveResult:
      properties:
        projectId:
          type: string
        removedStorageDir:
          type: boolean
      required:
      - projectId
      - removedStorageDir
      type: object
    SCMConfig:
      properties:
        package:
          type: string
        path:
          type: string
        plugin:
          type: string
        webhook:
          $ref: '#/components/schemas/SCMWebhookConfig'
      type: object
    SCMWebhookConfig:
      properties:
        deliveryHeader:
          type: string
        enabled:
          type:
          - "null"
          - boolean
        eventHeader:
          type: string
        maxBodyBytes:
          type: integer
        path:
          type: string
        secretEnvVar:
          type: string
        signatureHeader:
          type: string
      type: object
    Summary:
      properties:
        id:
          type: string
        name:
          type: string
        resolveError:
          description: Present iff the project is degraded.
          type: string
        sessionPrefix:
          type: string
      required:
      - id
      - name
      - sessionPrefix
      type: object
    TrackerConfig:
      properties:
        package:
          type: string
        path:
          type: string
        plugin:
          type: string
      type: object
    UpdateConfigInput:
      properties:
        agent:
          additionalProperties: {}
          type: object
        reactions:
          additionalProperties:
            $ref: '#/components/schemas/ReactionConfig'
          type:
          - "null"
          - object
        runtime:
          additionalProperties: {}
          type: object
        scm:
          $ref: '#/components/schemas/SCMConfig'
        tracker:
          $ref: '#/components/schemas/TrackerConfig'
      type: object
tags:
- description: Project registry, configuration, and lifecycle administration
  name: projects